Speed Boost Power Up
On collision with the object tag type "player" the player's speed will increase for a specific amount of time. On collision the Speed Boost Power Up will delete itself from the game scene.

Bugs - Destroy(gameObject) Destroys the code before the Invoke method. Meaning the speed up code is being executed and permanently set before the Invoke method is executed. (which it cant be executed as the object script is being deleted with the object).