Checkpoint System Implementation
Checkpoint Switch statement has been implemented, checkpoints cannot be skipped. The checkpoints are incremental, so an onTrigger has to Trigger every checkpoint before the finish line before the finishing line trigger is set to true.

Next: Time stopping on finish line collider to replicate the final time of the race.