Decrement Speed on Collision
Tester script implemented on collision it will decrement the speed of the player. For example: If the player went over grass the grass collider would decrement the players speed.

ThirdPersonCharacterMovement script (possible improvement) to the jumping to give it a more realistic feel. Increasing the gravity over Time.deltaTime to simulate a more realistic jump and fall.