Missile Tracking Player
Missile script implemented which will follow the players transform, this is simple code just to get things working which it is.

Bugs so far: Missile clipping through floor (when player is beneath ground).
Missile spins around the player (can be fixed by destroying itself on collision)

Want to be implemented later: Ability to at run-time give the player the ability to fire at targets that are only infront of the player, the object is hard-implemented atm.