PlayerHealth & Respawn Tweaks
PlayerHealth has been implemented, tested the collision on speed (more speed more damage delt to player). Implemented a "WallCollider" for testing purposes.

Missile does damage on the player now and destroys itself on collision.

When the players health is == 0 a IEnumerator of (1) second waits for one second before respawning back to nearest respawn point for later implementation of animations.