Rocket PowerUps/Speed PowerUps & UI changes
A lot of tweaks to power ups, speed boost power up is no longer a child of the parent object (No huge benefit, no need for it to be on the player waste of memory. Used booleans instead to check if the player has the powerup)

Rocket powerup might be a child of parent object, this way it can easily be instantiated on top when the player uses the missile (this will have to be  updated every frame).

Simple UI button which restarts the whole scene, including checkpoints/power ups and spawn locations.

Bugs: Timer does not stop when scene is reloaded.

Todo: Missile needs to track closest players, complete re-write will be needed for online play as hard-implementing objects for it to track is unrealistic.