Null movement on Respawn
If the player has less than or equal to 0 health, the player is dead. If the player is dead movement will be turned off for 2 seconds (during the respawn of the character), another Invoke 0.5 seconds after respawn the "isRespawning" boolean will be changed to false again, allowing movement as usual.